refactor(swiftui): split DocumentInspectorArtifactsTab into focused files (#1703)
